commands tests refactoring
