+++ test/suite/commands/index.test.ts
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 1998-2025 KX Systems Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the
+ * License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+import { readdirSync } from "fs";
+import { join } from "path";
+
+describe("Commands", () => {
+  const testFiles = readdirSync(__dirname)
+    .filter((file) => file.endsWith(".test.js") && file !== "index.test.js")
+    .sort((a, b) => a.localeCompare(b));
+
+  testFiles.forEach((file) => {
+    const filePath = join(__dirname, file);
+    eval("require")(filePath);
+  });
+});
